src/requirements.txt
defusedxml==0.4.1
Django==1.10.4
django-braces==1.10.0
django-filter==1.0.1
django-oauth-toolkit==0.11.0
django-rest==0.0.1
djangorestframework==3.5.3
djangorestframework-gis==0.11
mwoauth==0.2.8
oauth2-provider==0.0
oauthlib==2.0.1
packaging==16.8
pip-review==0.5.2
pluggy==0.4.0
psycopg2==2.6.2
py==1.4.32
PyJWT==1.4.2
pyparsing==2.2.0
pytest==3.0.5
python-social-auth==0.3.3
python3-openid==3.0.10
requests==2.12.5
requests-oauthlib==0.7.0
six==1.10.0
social-auth-app-django==0.1.0
social-auth-core==0.2.1
tox==2.5.0
virtualenv==15.1.0
